Fix issue #4
Checking for file existence and catching potential IO exceptions when
getting the length of a file.
Check that the file exists before getting it's length. Used GetFileSize
from IsolatedStorageDirectory as an example and also made that bullet
proof. Unit test to ensure that the method does not throw and exception,
it will just be ignored because the file will be less than "5MB".