Handles unauthorized requests globally
Refactors the handling of 401 errors to a central location to ensure consistent user authentication across the application.

This change centralizes the logic for handling 401 Unauthorized responses. Instead of checking for 401 errors in specific routes, it now checks in the main layout.

It also prevents unnecessary error propagation for expected 401 errors.

https://github.com/svecosystem/runed/issues/267#issuecomment-3055222286