Next: filtering improvements (#1959)
* next: Allow creating a custom keyword filter from current text.

* improved keyboard nav of string and number filters.

* improved nav of boolean filters

* Refactors faceted filter actions and builder

Simplifies the faceted filter actions component by using the Button component and Separator component instead of the Command component.

Improves the faceted filter builder by enhancing searchability by including builder titles as keywords.

Adds autofocus to the input field in dropdown and multi-select components.

* Adds default project filter to events views

Ensures that the project filter is always present in events views by default.

This change ensures that the events, issues and stream pages will always have a default project filter applied when the page loads.
It also sets a priority for the project builder in the faceted filter.

* Updates aria-describedby attribute for accessibility

Changes the aria-describedby attribute to reference the "-help" element
instead of the "-description" element for improved accessibility.

This ensures screen readers correctly associate the button with its
help text.