Adds stack events chart to stack card
Implements a stack events chart on the stack card to visualize event occurrences over the last 7 days.
This provides users with a quick overview of recent event activity related to a specific stack.
It includes handling for empty data sets and uses a skeleton loader while data is loading.

The changes also include a minor fix to faceted filter builder for date filters.