Adds row href to data tables
This change adds the ability to navigate to a specific URL when a row in a data table is clicked.

The `rowHref` prop is added to the `DataTable.Body` component, allowing a URL to be specified for each row. When a row is clicked, the browser will navigate to the specified URL.

Also adds default toolbar component to project and organization tables