Refactors faceted filter builders and updates filters
Moves the `gotoLogin` function to the bottom of the `index.svelte.ts` file for better organization.

Updates faceted filter builders to use getters for properties, avoiding the `state_referenced_locally` warning.

Adds a `$effect.pre` to faceted filter components to correctly initialize state.

Updates the projects query in `ProjectFacetedFilter` to reactively update when the organization changes.

Updates the chart id to be a derived value to update when the id changes.